@@ -11,7 +11,7 @@ license = "MIT"
 name = "napi-derive-ohos"
 readme = "README.md"
 repository = "https://github.com/ohos-rs/ohos-rs"
-rust-version = "1.78"
+rust-version = "1.80"
 version = "1.0.0-beta.7"
 
 [package.metadata.workspaces]
