@@ -11,7 +11,7 @@ license = "MIT"
 name = "napi-ohos"
 readme = "README.md"
 repository = "https://github.com/ohos-rs/ohos-rs"
-rust-version = "1.78"
+rust-version = "1.80"
 version = "1.0.0-beta.7"
 
 [lib]
@@ -63,9 +63,8 @@ dyn-symbols        = ["napi-sys-ohos/dyn-symbols"]
 ohos               = ["napi8", "napi-sys-ohos/ohos", "compat-mode"]
 
 [dependencies]
-bitflags  = "2"
-ctor      = "0.2"
-once_cell = "1.19"
+bitflags = "2"
+ctor     = "0.2"
 
 [dependencies.anyhow]
 optional = true
